requirements.txt
altgraph==0.17.4
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
emoji==2.15.0
et_xmlfile==2.0.0
filelock==3.16.1
flake8==7.2.0
Flask==3.1.2
fonttools==4.55.8
fuzzy-search==2.5.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
googleapis-common-protos==1.71.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
iniconfig==2.1.0
instaloader==4.15
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Levenshtein==0.27.1
MarkupSafe==3.0.3
matplotlib==3.10.0
mccabe==0.7.0
MiniSom==2.3.5
nltk==3.9.1
numpy==2.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pbr==6.1.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.23
pyflakes==3.3.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.1
pytz==2025.1
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
regex==2024.11.6
reportlab==4.4.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.38.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stevedore==5.4.0
thefuzz==0.22.1
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.1
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.5.0
vaderSentiment==3.3.2
virtualenv==20.29.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.1
